pyproject.toml
[tool.poetry]
name = "strava-to-fittrackee"
version = "0.1.0"
description = "Pull workouts from Strava and upload to FitTrackee"
authors = ["Joshua Taillon <jat255@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "strava_to_fittrackee"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
requests-oauthlib = "^1.3.1"
python-dotenv = "^0.21.0"
gpxpy = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
